Convert TSNetworkManager to protocol
This makes our (broken) tests less brittle

Note that this is a much simpler process where networkManager was being
treated as an injected dependency...

Translating this for class methods, and singleton access (and
particularly singleton access *within* a class method) is harder to
properly translate.

In pursuit of a minimal change, I simply updated the class names. Given
infinite time to verify the changes, I'd prefer to convert class methods
that access external singletons to instance methods that access injected
dependencies.

// FREEBIE